refactor(utils): privatize some utils
